Progress unifying tables
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS29]$/Almonaster;C8810

+++ Main/SqlDatabaseBridge/Utils.cpp
@@ -12,7 +12,7 @@ System::Object^ Convert(const Variant& v)
     case V_INT64:
         return v.GetInteger64();
     case V_FLOAT:
-        return v.GetFloat();
+        return (double)v.GetFloat();
     case V_STRING:
         return gcnew System::String(v.GetString());
     default:
